+++ Tests/StreamChatTests/Database/DTOs/MessageDTO_Tests.swift
@@ -2327,6 +2327,51 @@ final class MessageDTO_Tests: XCTestCase {
         XCTAssertEqual(retrievedMessages.filter { msg in msg.id == message.id }.count, 1)
     }
 
+    func test_allAttachmentsAreUploadedOrEmptyPredicate_whenAllUploadedFromServer_returnsMessages() throws {
+        let request = NSFetchRequest<MessageDTO>(entityName: MessageDTO.entityName)
+        request.sortDescriptors = [NSSortDescriptor(keyPath: \MessageDTO.defaultSortingKey, ascending: true)]
+        request.predicate = MessageDTO.allAttachmentsAreUploadedOrEmptyPredicate()
+
+        let cid = ChannelId.unique
+        let message: MessagePayload =
+            .dummy(
+                type: .regular,
+                messageId: .unique,
+                attachments: [],
+                authorUserId: .unique,
+                createdAt: Date(timeIntervalSince1970: 1),
+                channel: .dummy(cid: cid)
+            )
+
+        try database.writeSynchronously { session in
+            let message = try session.saveMessage(
+                payload: message,
+                for: .unique,
+                syncOwnReactions: true,
+                cache: nil
+            )
+
+            let attachment1 = try session.saveAttachment(
+                payload: .image(),
+                id: .init(cid: cid, messageId: message.id, index: 1)
+            )
+            attachment1.localState = nil
+
+            let attachment2 = try session.saveAttachment(
+                payload: .image(),
+                id: .init(cid: cid, messageId: message.id, index: 2)
+            )
+            attachment2.localState = nil
+
+            message.attachments.insert(attachment1)
+            message.attachments.insert(attachment2)
+        }
+
+        var retrievedMessages: [MessageDTO] = []
+        retrievedMessages = try database.viewContext.fetch(request)
+        XCTAssertEqual(retrievedMessages.filter { msg in msg.id == message.id }.count, 1)
+    }
+
     func test_allAttachmentsAreUploadedOrEmptyPredicate_whenSomeUploaded_shouldNotReturnMessages() throws {
         let request = NSFetchRequest<MessageDTO>(entityName: MessageDTO.entityName)
         request.sortDescriptors = [NSSortDescriptor(keyPath: \MessageDTO.defaultSortingKey, ascending: true)]